refactor: move lockfile.rs to args module (#16818)
This should be in the `args` folder as it's similar to `config_file`.
