+++ cli/args/mod.rs
@@ -1,7 +1,8 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 
-pub mod config_file;
-pub mod flags;
+mod config_file;
+mod flags;
+mod lockfile;
 
 mod flags_allow_net;
 
@@ -10,6 +11,8 @@ pub use config_file::ConfigFile;
 pub use config_file::EmitConfigOptions;
 pub use config_file::FmtConfig;
 pub use config_file::FmtOptionsConfig;
+pub use config_file::IgnoredCompilerOptions;
+pub use config_file::JsxImportSourceConfig;
 pub use config_file::LintConfig;
 pub use config_file::LintRulesConfig;
 pub use config_file::MaybeImportsResult;
@@ -17,6 +20,8 @@ pub use config_file::ProseWrap;
 pub use config_file::TestConfig;
 pub use config_file::TsConfig;
 pub use flags::*;
+pub use lockfile::Lockfile;
+pub use lockfile::LockfileError;
 
 use deno_ast::ModuleSpecifier;
 use deno_core::anyhow::anyhow;
@@ -36,7 +41,6 @@ use std::net::SocketAddr;
 use std::path::PathBuf;
 use std::sync::Arc;
 
-use crate::args::config_file::JsxImportSourceConfig;
 use crate::deno_dir::DenoDir;
 use crate::emit::get_ts_config_for_emit;
 use crate::emit::TsConfigType;
@@ -45,7 +49,6 @@ use crate::emit::TsTypeLib;
 use crate::file_fetcher::get_root_cert_store;
 use crate::file_fetcher::CacheSetting;
 use crate::fs_util;
-use crate::lockfile::Lockfile;
 use crate::version;
 
 /// Overrides for the options below that when set will
